Add VS2022 for Sonnet Stable 1.3 (initial release Nov 2020) Using current projects from current master as much as possible. Minor functional change to Objective (get Expression) and using latest Sonnet tests, including Native. The aim is to release 1.3.1 based on the latest Cbc 2.10, where the current 1.3.0 is based on Cbc 2.10.5.
