+++ .coin-or/Dependencies
@@ -1,0 +1,7 @@
+Data/Sample  https://github.com/coin-or-tools/Data-Sample stable/1.2
+Data/miplib3  https://github.com/coin-or-tools/Data-miplib3 stable/1.2
+CoinUtils  https://github.com/coin-or/CoinUtils stable/2.11
+Osi  https://github.com/coin-or/Osi stable/0.108
+Clp  https://github.com/coin-or/Clp stable/1.17
+Cgl  https://github.com/coin-or/Cgl stable/0.60
+Cbc  https://github.com/coin-or/Cbc stable/2.10
