+++ .gitignore
@@ -4,22 +4,27 @@
 
 *.suo
 *.db
-/.vs
-/MSVisualStudio/v10/.vs
-/MSVisualStudio/v10/Sonnet/obj
-/MSVisualStudio/v10/SonnetExamples/obj
-/MSVisualStudio/v10/SonnetTest/obj
-/MSVisualStudio/v16/.vs
-/MSVisualStudio/v16/Sonnet/bin
-/MSVisualStudio/v16/Sonnet/obj
-/MSVisualStudio/v16/SonnetExamples/bin
-/MSVisualStudio/v16/SonnetExamples/obj
-/MSVisualStudio/v16/SonnetTest/bin
-/MSVisualStudio/v16/SonnetTest/obj
-/MSVisualStudio/v16/SonnetWrapper/bin
-/MSVisualStudio/v16/SonnetWrapper/obj
-/MSVisualStudio/v16/x64*
-/MSVisualStudio/v16/Win32*
 *.bak
 *.user
-*.user
+app.aps
+
+**/.vs/
+**/.sonarlint/
+
+#ignore build folders
+/MSVisualStudio/**/obj/
+/MSVisualStudio/**/bin/
+/MSVisualStudio/**/Release/
+/MSVisualStudio/**/ReleaseParallel/
+/MSVisualStudio/**/Debug/
+/MSVisualStudio/**/x64*/
+/MSVisualStudio/**/Win32*/
+
+/MSVisualStudio/v16/packages/
+/MSVisualStudio/v16/TestResults/
+/MSVisualStudio/v16/UpgradeLog.htm
+/MSVisualStudio/v16/Sonnet/Properties/launchSettings.json
+/MSVisualStudio/v16/SonnetExamples/Properties/launchSettings.json
+/MSVisualStudio/v16/SonnetTest/Properties/launchSettings.json
+
+/MSVisualStudio/v17/packages/
