+++ MSVisualStudio/v17/SonnetExamples/SonnetExamples.csproj
@@ -1,0 +1,31 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+    <TargetFramework>net6.0</TargetFramework>
+    <Platforms>x64;x86</Platforms>
+    <CodeAnalysisRuleSet>..\.sonarlint\sonnet-master-cicsharp.ruleset</CodeAnalysisRuleSet>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="..\..\..\examples\Example1.cs" Link="Example1.cs" />
+    <Compile Include="..\..\..\examples\Example2.cs" Link="Example2.cs" />
+    <Compile Include="..\..\..\examples\Example3.cs" Link="Example3.cs" />
+    <Compile Include="..\..\..\examples\Example4.cs" Link="Example4.cs" />
+    <Compile Include="..\..\..\examples\Example5.cs" Link="Example5.cs" />
+    <Compile Include="..\..\..\examples\Example6.cs" Link="Example6.cs" />
+    <Compile Include="..\..\..\examples\Example6b.cs" Link="Example6b.cs" />
+    <Compile Include="..\..\..\examples\Program.cs" Link="Program.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <AdditionalFiles Include="..\.sonarlint\sonnet-master-ci\CSharp\SonarLint.xml" Link="SonarLint.xml" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\SonnetWrapper\SonnetWrapper.vcxproj" />
+    <ProjectReference Include="..\Sonnet\Sonnet.csproj" />
+  </ItemGroup>
+
+</Project>
