+++ MSVisualStudio/v17/SonnetTestNET4/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MSTest.TestAdapter" version="2.2.10" targetFramework="net48" />
+  <package id="MSTest.TestFramework" version="2.2.10" targetFramework="net48" />
+</packages>
