+++ MSVisualStudio/v17/SonnetWrapperTest/newlengths.mps
@@ -1,0 +1,203 @@
+NAME          ClpDefau  FREE
+ROWS
+ N  OBJROW
+ L  order_total
+ G  group1min
+ L  group1max
+ G  group1nmin
+ L  group1nmax
+ L  iflength1_L300
+ L  iflength1_L330
+ L  iflength1_L360
+ L  iflength1_L390
+ L  iflength1_L420
+ L  iflength1_L450
+ L  iflength1_L480
+ L  iflength1_L510
+ L  iflength1_L540
+ L  iflength1_L570
+ L  iflength1_L600
+ G  iflength2_L300
+ G  iflength2_L330
+ G  iflength2_L360
+ G  iflength2_L390
+ G  iflength2_L420
+ G  iflength2_L450
+ G  iflength2_L480
+ G  iflength2_L510
+ G  iflength2_L540
+ G  iflength2_L570
+ G  iflength2_L600
+ G  mingroup1single_case_n_out_m_L420
+ G  mingroup1single_case_n_out_m_L450
+ G  mingroup1single_case_n_out_m_L480
+ L  maxgroup1single_case_n_out_m_L420
+ L  maxgroup1single_case_n_out_m_L450
+ L  maxgroup1single_case_n_out_m_L480
+ G  mingroup1single_L420
+ G  mingroup1single_L450
+ G  mingroup1single_L480
+ L  maxgroup1single_L420
+ L  maxgroup1single_L450
+ L  maxgroup1single_L480
+COLUMNS
+    x_L300 OBJROW 350.        
+    x_L300 order_total 3.5         
+    x_L300 iflength1_L300 1.          
+    x_L300 iflength2_L300 1.          
+    x_L330 order_total 3.8         
+    x_L330 iflength1_L330 1.          
+    x_L330 iflength2_L330 1.          
+    x_L360 OBJROW  -420.      
+    x_L360 order_total 4.2         
+    x_L360 iflength1_L360 1.          
+    x_L360 iflength2_L360 1.          
+    x_L390 OBJROW 1564.       
+    x_L390 order_total 4.6         
+    x_L390 iflength1_L390 1.          
+    x_L390 iflength2_L390 1.          
+    x_L420 OBJROW  -2100.     
+    x_L420 order_total 5.          
+    x_L420 group1min 5.          
+    x_L420 group1max 5.          
+    x_L420 iflength1_L420 1.          
+    x_L420 iflength2_L420 1.          
+    x_L420 mingroup1single_case_n_out_m_L420 4.75        
+    x_L420 mingroup1single_case_n_out_m_L450  -0.25      
+    x_L420 mingroup1single_case_n_out_m_L480  -0.25      
+    x_L420 maxgroup1single_case_n_out_m_L450  -5.        
+    x_L420 maxgroup1single_case_n_out_m_L480  -5.        
+    x_L420 mingroup1single_L420 5.          
+    x_L420 maxgroup1single_L450  -5.        
+    x_L420 maxgroup1single_L480  -5.        
+    x_L450 OBJROW  -2650.     
+    x_L450 order_total 5.3         
+    x_L450 group1min 5.3         
+    x_L450 group1max 5.3         
+    x_L450 iflength1_L450 1.          
+    x_L450 iflength2_L450 1.          
+    x_L450 mingroup1single_case_n_out_m_L420  -0.265     
+    x_L450 mingroup1single_case_n_out_m_L450 5.035       
+    x_L450 mingroup1single_case_n_out_m_L480  -0.265     
+    x_L450 maxgroup1single_case_n_out_m_L420  -5.3       
+    x_L450 maxgroup1single_case_n_out_m_L480  -5.3       
+    x_L450 mingroup1single_L450 5.3         
+    x_L450 maxgroup1single_L420  -5.3       
+    x_L450 maxgroup1single_L480  -5.3       
+    x_L480 OBJROW 342.        
+    x_L480 order_total 5.7         
+    x_L480 group1min 5.7         
+    x_L480 group1max 5.7         
+    x_L480 iflength1_L480 1.          
+    x_L480 iflength2_L480 1.          
+    x_L480 mingroup1single_case_n_out_m_L420  -0.285     
+    x_L480 mingroup1single_case_n_out_m_L450  -0.285     
+    x_L480 mingroup1single_case_n_out_m_L480 5.415       
+    x_L480 maxgroup1single_case_n_out_m_L420  -5.7       
+    x_L480 maxgroup1single_case_n_out_m_L450  -5.7       
+    x_L480 mingroup1single_L480 5.7         
+    x_L480 maxgroup1single_L420  -5.7       
+    x_L480 maxgroup1single_L450  -5.7       
+    x_L510 OBJROW 720.        
+    x_L510 order_total 6.          
+    x_L510 iflength1_L510 1.          
+    x_L510 iflength2_L510 1.          
+    x_L540 OBJROW 1260.       
+    x_L540 order_total 6.3         
+    x_L540 iflength1_L540 1.          
+    x_L540 iflength2_L540 1.          
+    x_L570 OBJROW  -2376.     
+    x_L570 order_total 6.6         
+    x_L570 iflength1_L570 1.          
+    x_L570 iflength2_L570 1.          
+    x_L600 OBJROW  -1400.     
+    x_L600 order_total 7.          
+    x_L600 iflength1_L600 1.          
+    x_L600 iflength2_L600 1.          
+    n_L420 group1nmin 1.          
+    n_L420 group1nmax 1.          
+    n_L420 iflength1_L420  -300000.   
+    n_L420 iflength2_L420  -300000.   
+    n_L420 mingroup1single_case_n_out_m_L420  -300000.   
+    n_L450 group1nmin 1.          
+    n_L450 group1nmax 1.          
+    n_L450 iflength1_L450  -300000.   
+    n_L450 iflength2_L450  -300000.   
+    n_L450 mingroup1single_case_n_out_m_L450  -300000.   
+    n_L480 group1nmin 1.          
+    n_L480 group1nmax 1.          
+    n_L480 iflength1_L480  -300000.   
+    n_L480 iflength2_L480  -300000.   
+    n_L480 mingroup1single_case_n_out_m_L480  -300000.   
+    n_L300 iflength1_L300  -300000.   
+    n_L300 iflength2_L300  -300000.   
+    n_L330 iflength1_L330  -300000.   
+    n_L330 iflength2_L330  -300000.   
+    n_L360 iflength1_L360  -300000.   
+    n_L360 iflength2_L360  -300000.   
+    n_L390 iflength1_L390  -300000.   
+    n_L390 iflength2_L390  -300000.   
+    n_L510 iflength1_L510  -300000.   
+    n_L510 iflength2_L510  -300000.   
+    n_L540 iflength1_L540  -300000.   
+    n_L540 iflength2_L540  -300000.   
+    n_L570 iflength1_L570  -300000.   
+    n_L570 iflength2_L570  -300000.   
+    n_L600 iflength1_L600  -300000.   
+    n_L600 iflength2_L600  -300000.   
+RHS
+    RHS order_total 100.        
+    RHS group1min 30.         
+    RHS group1max 70.         
+    RHS group1nmin 3.          
+    RHS group1nmax 3.          
+    RHS iflength2_L300  -300000.   
+    RHS iflength2_L330  -300000.   
+    RHS iflength2_L360  -300000.   
+    RHS iflength2_L390  -300000.   
+    RHS iflength2_L420  -300000.   
+    RHS iflength2_L450  -300000.   
+    RHS iflength2_L480  -300000.   
+    RHS iflength2_L510  -300000.   
+    RHS iflength2_L540  -300000.   
+    RHS iflength2_L570  -300000.   
+    RHS iflength2_L600  -300000.   
+    RHS mingroup1single_case_n_out_m_L420  -300000.   
+    RHS mingroup1single_case_n_out_m_L450  -300000.   
+    RHS mingroup1single_case_n_out_m_L480  -300000.   
+BOUNDS
+ UI BOUND x_L300        1e+30
+ UI BOUND x_L330        1e+30
+ UI BOUND x_L360        1e+30
+ UI BOUND x_L390        1e+30
+ UI BOUND x_L420        1e+30
+ UI BOUND x_L450        1e+30
+ UI BOUND x_L480        1e+30
+ UI BOUND x_L510        1e+30
+ UI BOUND x_L540        1e+30
+ UI BOUND x_L570        1e+30
+ UI BOUND x_L600        1e+30
+ BV BOUND n_L420 1.          
+ BV BOUND n_L450 1.          
+ BV BOUND n_L480 1.          
+ BV BOUND n_L300 1.          
+ BV BOUND n_L330 1.          
+ BV BOUND n_L360 1.          
+ BV BOUND n_L390 1.          
+ BV BOUND n_L510 1.          
+ BV BOUND n_L540 1.          
+ BV BOUND n_L570 1.          
+ BV BOUND n_L600 1.          
+QUADOBJ
+    x_L300 x_L300 24.5        
+    x_L330 x_L330 28.88       
+    x_L360 x_L360 35.28       
+    x_L390 x_L390 42.32       
+    x_L420 x_L420 50.         
+    x_L450 x_L450 56.18       
+    x_L480 x_L480 64.98       
+    x_L510 x_L510 72.         
+    x_L540 x_L540 79.38       
+    x_L570 x_L570 87.12       
+    x_L600 x_L600 98.         
+ENDATA
