+++ src/SonnetWrapper/NativeTests.cpp
@@ -1,0 +1,4 @@
+// Copyright (C) Jan-Willem Goossens 
+// This code is licensed under the terms of the Eclipse Public License (EPL).
+
+#include "NativeTests.h"
