+++ test/SonnetTest/AssemblyInfo.cs
@@ -1,33 +1,23 @@
-// Copyright (C) 2011, Jan-Willem Goossens 
-// All Rights Reserved.
+// Copyright (C) Jan-Willem Goossens 
 // This code is licensed under the terms of the Eclipse Public License (EPL).
 
 using System.Reflection;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
-//
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-//
 [assembly: AssemblyTitle("SonnetTest")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
+[assembly: AssemblyCompany("See https://github.com/coin-or/sonnet")]
 [assembly: AssemblyProduct("SonnetTest")]
-[assembly: AssemblyCopyright("Copyright ©  2011-2015")]
-[assembly: AssemblyTrademark("This code is licensed under the terms of the Eclipse Public License (EPL).")]
-[assembly: AssemblyCulture("")]		
+[assembly: AssemblyCopyright("This code is licensed under the terms of the Eclipse Public License v2.0 (EPL-2.0).")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
 
-//
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Revision and Build Numbers 
-// by using the '*' as shown below:
-[assembly: AssemblyVersion("1.1.1.0")]
-[assembly: AssemblyFileVersion("1.1.1.1")]
+[assembly: ComVisible(false)]
+
+[assembly: Guid("252e3e95-6e97-45e9-9470-26e11644c879")]
+
+// [assembly: AssemblyVersion("1.0.*")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
